# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Drakor & Webtoon List API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MediaItem {
  id: number;
  title: string;
  /** drakor | webtoon | short-dracin | indo */
  category: string;
  /** @nullable */
  genre?: string | null;
  /** plan-to-watch | watching | completed | dropped */
  status: string;
  /**
   * 1-10
   * @nullable
   */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  totalEpisodes?: number | null;
  /** @nullable */
  currentEpisode?: number | null;
  createdAt: string;
}

export interface MediaInput {
  /** @minLength 1 */
  title: string;
  category: string;
  genre?: string;
  status: string;
  rating?: number;
  notes?: string;
  totalEpisodes?: number;
  currentEpisode?: number;
}

export interface MediaUpdate {
  /** @minLength 1 */
  title?: string;
  category?: string;
  genre?: string;
  status?: string;
  rating?: number;
  notes?: string;
  totalEpisodes?: number;
  currentEpisode?: number;
}

export interface CategoryStats {
  category: string;
  total: number;
  completed: number;
  watching: number;
  planToWatch: number;
  dropped: number;
}

export interface MediaStats {
  totalAll: number;
  categories: CategoryStats[];
}

export type ListMediaParams = {
  category?: string;
  genre?: string;
  status?: string;
  search?: string;
};