# lib/api-zod/src/generated/types/categoryStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Drakor & Webtoon List API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryStats {
  category: string;
  total: number;
  completed: number;
  watching: number;
  planToWatch: number;
  dropped: number;
}