# lib/api-zod/src/generated/types/mediaItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Drakor & Webtoon List API
 * OpenAPI spec version: 0.1.0
 */

export interface MediaItem {
  id: number;
  title: string;
  /** drakor | webtoon | short-dracin | indo */
  category: string;
  /** @nullable */
  genre?: string | null;
  /** plan-to-watch | watching | completed | dropped */
  status: string;
  /**
   * 1-10
   * @nullable
   */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  totalEpisodes?: number | null;
  /** @nullable */
  currentEpisode?: number | null;
  createdAt: string;
}