# lib/api-zod/src/generated/types/mediaStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Drakor & Webtoon List API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryStats } from "./categoryStats";

export interface MediaStats {
  totalAll: number;
  categories: CategoryStats[];
}