# lib/api-zod/src/generated/types/mediaUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Drakor & Webtoon List API
 * OpenAPI spec version: 0.1.0
 */

export interface MediaUpdate {
  /** @minLength 1 */
  title?: string;
  category?: string;
  genre?: string;
  status?: string;
  rating?: number;
  notes?: string;
  totalEpisodes?: number;
  currentEpisode?: number;
}